pyproject.toml
[project]
name = "agentischer-docstore"
version = "1.0.0"
description = "Agentischer Document Store – Akte & WissensDB Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "aiofiles>=23.2.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "striprtf>=0.0.26",
    "markdown>=3.5",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10", "pdf2image>=1.17.0"]
vectors = ["qdrant-client>=1.7.0", "sentence-transformers>=2.3.0"]
queue = ["celery[redis]>=5.3.0"]
